requirements.txt
Flask==3.0.0
gunicorn
openai>=1.0.0
pdfplumber
python-docx
reportlab
